商品详情布局写好了，上拉刷新，下拉刷新写好了，就是出不来

+++ src/cn/ucai/fulicenter/bean/NewGoodBean.java
@@ -49,7 +49,7 @@ public class NewGoodBean  implements Serializable {
     private String colorName;
     private String colorCode;
     private String colorUrl;
-    private int addTime;
+    private long addTime;
 
     public int getId() {
         return id;
@@ -187,11 +187,11 @@ public class NewGoodBean  implements Serializable {
         this.colorUrl = colorUrl;
     }
 
-    public int getAddTime() {
+    public long getAddTime() {
         return addTime;
     }
 
-    public void setAddTime(int addTime) {
+    public void setAddTime(long addTime) {
         this.addTime = addTime;
     }
